collect rollout submodules
