+++ malib/rollout/__init__.py
@@ -22,6 +22,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from .rollout_worker import RolloutWorker
-from .pipe_inference.client import InferenceClient
-from .pipe_inference.server import InferenceWorkerSet
+from .pb_rolloutworker import RolloutWorker
+from .inference.pipe.client import InferenceClient
+from .inference.pipe.server import InferenceWorkerSet
