+++ malib/rollout/tests/test_inference.py
@@ -9,8 +9,8 @@ from malib import settings
 from malib.common.strategy_spec import StrategySpec
 from malib.algorithm.random import RandomPolicy
 from malib.rollout.envs import dummy_env
-from malib.rollout.pipe_inference.client import InferenceClient
-from malib.rollout.pipe_inference.server import InferenceWorkerSet
+from malib.rollout.inference.pipe.client import InferenceClient
+from malib.rollout.inference.pipe.server import InferenceWorkerSet
 from malib.backend.offline_dataset_server import OfflineDataset
 from malib.backend.parameter_server import ParameterServer
