+++ setup.py
@@ -18,8 +18,8 @@ setup(
     python_requires=">=3.7",
     install_requires=[
         "wrapt",
-        "ray==1.8.0",
-        "pettingzoo==1.13.1",
+        "ray==1.13.0",
+        "pettingzoo==1.19.0",
         "grpcio-tools",
         "protobuf3-to-dict",
         "pickle5",
@@ -31,7 +31,7 @@ setup(
         "pymongo",
         "psutil",
         "pyecharts",
-        "open_spiel==1.0.2",
+        "open_spiel>=1.0.2",
         "supersuit==3.3.1",
         "multi-agent-ale-py==0.1.11",
         "autorom==0.4.2",
